Make blocking I/O operations interruptible.
Blocking read, write and accept calls are now interrupted if
the fiber is interrupted. This is done by asynchronously
closing the channel on fiber interrupt.

The only other to way to interrupt blocking calls in a
NIO channel is to interrupt (in the Java sense) the
blocked thread. However this also results in the channel
being closed and requires more overhead on the ZIO side.