Add more efficient blocking API.
To correctly encapsulate a blocking NIO call in a ZIO effect value
we need two things:
 - switch the fiber to a thread from the blocking pool, to avoid
 consuming a thread from the fixed-size default pool
 - fork and join the effect to allow attaching interrupt handling

Doing the above for each blocking read or write call imposes
significant overhead in many situations.

The new blocking APIs introduced here allow performing the
blocking setup once for a whole sequence of read/write effects
(which may be looping/recursive).

`BlockingPerformance.scala` quantifies the difference in the
case of writing to a `FileChannel`, while demonstrating use
of the new API.